co3d_3d/requirements.txt
# Pytorch and lightning
pytorch_lightning==1.5.5

# configuration and logging
gin_config
pandas
wandb
test_tube

# Data loading
plyfile
retrying
scipy
matplotlib
h5py
scikit-learn

# Utility
tqdm
ninja
